ArchiveBase.cs: Using the `handler` variable instead of `FileAdded`.
`FileAdded` is copied locally to check it for null before using it.
Going back to using `FileAdded` after checking the local variable
defeats the purpose and can result in a NullReferenceError if something
nulls it out between check and use. (TOCTOU)

PVS-Studio V3083 (Med): Unsafe invocation of event 'FileAdded'.
NullReferenceExeception is possible. Consider assigning event to a local
variable before invoking it.